pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hum"
version = "0.1.0"
description = "Small agents that keep running."
requires-python = ">=3.11"
dependencies = [
    "agent-client-protocol",
    "aiohttp>=3.9.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.scripts]
hum = "hum.cli:main"
